refactor: Move pushup tracking features into own folders
