+++ lib/features/pushup_tracking/pushup_tracking.dart
@@ -8,9 +8,9 @@ import 'package:pushup_bro/core/model/pushup_set.dart';
 import 'package:pushup_bro/features/pushup_tracking/cubit/airpods_tracker_cubit.dart';
 import 'package:pushup_bro/features/pushup_tracking/cubit/airpods_tracker_state.dart';
 import 'package:pushup_bro/features/pushup_tracking/cubit/pushup_cubit.dart';
-import 'package:pushup_bro/features/pushup_tracking/widgets/finished_set_bottom_sheet.dart';
+import 'package:pushup_bro/features/pushup_tracking/widgets/finished_pushup_sheet/finished_set_bottom_sheet.dart';
 import 'package:pushup_bro/features/pushup_tracking/widgets/monkey.dart';
-import 'package:pushup_bro/features/pushup_tracking/widgets/news_banner.dart';
+import 'package:pushup_bro/features/pushup_tracking/widgets/news_banner/news_banner.dart';
 import 'package:pushup_bro/features/pushup_tracking/widgets/pushup_counter.dart';
 import 'package:pushup_bro/features/pushup_tracking/widgets/pushup_tracking_bottom_row.dart';
 import 'package:pushup_bro/features/pushup_tracking/widgets/pushup_tracking_top_row.dart';
