+++ lib/features/pushup_tracking/widgets/item_menu/item_menu.dart
@@ -4,7 +4,7 @@ import 'package:carbon_icons/carbon_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:pushup_bro/core/extensions/build_context_ext.dart';
 import 'package:pushup_bro/core/model/booster_items.dart';
-import 'package:pushup_bro/features/pushup_tracking/widgets/item_menu_item.dart';
+import 'package:pushup_bro/features/pushup_tracking/widgets/item_menu/item_menu_item.dart';
 
 class ItemMenu extends StatelessWidget {
   const ItemMenu({super.key});
